Avoided a deprecation warning

+++ openquake/hazardlib/geo/surface/base.py
@@ -505,10 +505,10 @@ class BaseQuadrilateralSurface(BaseSurface):
 
         total_len_y = (len(mesh.depths) - 1) * mesh_spacing
         y_distance = hypo_loc[1] * total_len_y
-        y_node = numpy.round(y_distance / mesh_spacing)
+        y_node = int(numpy.round(y_distance / mesh_spacing))
         total_len_x = (len(mesh.lons[y_node]) - 1) * mesh_spacing
         x_distance = hypo_loc[0] * total_len_x
-        x_node = numpy.round(x_distance / mesh_spacing)
+        x_node = int(numpy.round(x_distance / mesh_spacing))
         hypocentre = Point(mesh.lons[y_node][x_node],
                            mesh.lats[y_node][x_node],
                            mesh.depths[y_node][x_node])
